requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
Flask-Caching==2.1.0
psycopg2-binary==2.9.9
APScheduler==3.10.1
requests-cache==1.1.0
Flask-SocketIO==5.3.4
python-socketio==5.8.0
